pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "phone2app"
version = "0.1.0"
description = "Android real-device app latency and performance analysis toolkit."
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
appium = [
    "Appium-Python-Client>=5.0",
    "selenium>=4.0",
]
uiautomator2 = [
    "uiautomator2>=3.0",
]
test = [
    "pytest>=7.0",
]

[project.scripts]
phone2app = "phone2app.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "e2e: real-device Appium tests",
]
